+++ paddle/fluid/pybind/eager.cc
@@ -17,7 +17,7 @@ limitations under the License. */
 #include "paddle/fluid/eager/api/api.h"
 #include "paddle/fluid/eager/autograd_meta.h"
 #include "paddle/fluid/eager/function_api.h"
-#include "paddle/fluid/eager/generated2/dygraph_forward_api.h"
+#include "paddle/fluid/eager/generated/dygraph_forward_api.h"
 #include "paddle/fluid/memory/allocation/allocator.h"
 #include "paddle/fluid/memory/memcpy.h"
 #include "paddle/fluid/platform/enforce.h"
